pyproject.toml
[tool.poetry]
name = "fridge-scan"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.112.0"
pillow = "^10.4.0"
uvicorn = "^0.30.5"
python-multipart = "^0.0.9"
jinja2 = "^3.1.4"
inference-sdk = "^0.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
